Neural updates, ComboSess plans
- Updated raster and PSTH to be more robust to data dimensions, add
extra info to graphs. processPSTH not finished - need to save graphs,
data.
- Added stim checks to BehavAnalysis - can be used to check
seeds/rates/type etc where stim are expected to be the same. Not
finished yet, will need testing when more data available.
- Added ComboSess plans - ComboSess will be equiv to Sess, handled by
sessions. Might need to alter Sess properties slightly as planning to
inherit in to CombSess.
- NeuralAnalysis and NeualPP are not specific to this analysis. Neural
is, but only due to requirement of behavioural data for epoching. It
might be worth simplifying this to require only trial times later, so
Neural can be reused for spatial analyis. Sess objects are specific to
this analysis because they handle behavioural analysis and analysis
linking behavioural and neural data.